refactor: 인프라 구현 분리
