+++ nolzo-core/src/main/java/com/noljo/nolzo/review/repository/ReviewRepository.java
@@ -1,12 +1,12 @@
 package com.noljo.nolzo.review.repository;
 
 import com.noljo.nolzo.review.entity.Review;
-import io.lettuce.core.dynamic.annotation.Param;
 import java.util.List;
+import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.Optional;
+import org.springframework.data.repository.query.Param;
 import org.springframework.data.jpa.repository.Query;
 
 public interface ReviewRepository extends JpaRepository<Review, Long> {
