+++ nolzo-infra/src/main/java/com/noljo/nolzo/global/aop/idempotent/AspectUtils.java
@@ -1,6 +1,7 @@
-package com.noljo.nolzo.global.aop.utils;
+package com.noljo.nolzo.global.aop.idempotent;
 
 import java.util.Optional;
+
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.reflect.MethodSignature;
 import org.springframework.expression.spel.standard.SpelExpressionParser;
